refactor(auth): moved pass validation to commons
