+++ bins/auth/src/core/services/error.rs
@@ -1,12 +1,9 @@
 #![allow(dead_code)]
-use super::password_validator::PasswordValidationErrors;
 use password_hash::errors::Error as PasswordHashError;
 use thiserror::Error as TError;
 
 #[derive(Debug, Clone, TError, PartialEq)]
 pub enum Error {
-    #[error("{}", .0)]
-    PasswordValidation(PasswordValidationErrors),
     #[error("{0}")]
     PasswordHash(PasswordHashError),
 }
